extract tabs into a dedicated module
