+++ www/front_src/src/Resources/Details/Body/index.tsx
@@ -1,15 +1,10 @@
 import * as React from 'react';
 
-import { isNil } from 'ramda';
-
 import { Tabs, Tab, makeStyles, AppBar } from '@material-ui/core';
 
-import { labelDetails, labelGraph } from '../../translatedLabels';
 import { DetailsSectionProps } from '..';
-import GraphTab from './tabs/Graph';
-import DetailsTab from './tabs/Details';
-import { ResourceDetails } from '../models';
-import { GraphEndpoints, TabEndpoints } from './models';
+import { TabEndpoints } from './models';
+import { TabById, tabs } from './tabs';
 
 const useStyles = makeStyles((theme) => {
   return {
@@ -34,38 +29,6 @@ const useStyles = makeStyles((theme) => {
     },
   };
 });
-
-const tabs = [
-  {
-    key: 0,
-    Component: DetailsTab,
-    title: labelDetails,
-    visible: (): boolean => true,
-  },
-  {
-    key: 1,
-    Component: GraphTab,
-    title: labelGraph,
-    visible: ({ statusGraph, performanceGraph }: GraphEndpoints): boolean =>
-      !isNil(performanceGraph) || !isNil(statusGraph),
-  },
-];
-
-interface TabByIdProps {
-  details?: ResourceDetails;
-  id: number;
-  endpoints: TabEndpoints;
-}
-
-const TabById = ({
-  id,
-  details,
-  endpoints,
-}: TabByIdProps): JSX.Element | null => {
-  const { Component } = tabs[id];
-
-  return <Component details={details} endpoints={endpoints} />;
-};
 
 type Props = {
   endpoints: TabEndpoints;
@@ -97,8 +60,8 @@ const Body = ({ details, endpoints, openTabId }: Props): JSX.Element => {
         >
           {tabs
             .filter(({ visible }) => visible(endpoints))
-            .map(({ key, title }) => (
-              <Tab key={key} label={title} disabled={details === undefined} />
+            .map(({ id, title }) => (
+              <Tab key={id} label={title} disabled={details === undefined} />
             ))}
         </Tabs>
       </AppBar>
